feat: More stylish edit/create pages
