+++ assets/css/_datepicker.scss
@@ -11,7 +11,6 @@
 
 .m-datepicker__icon {
   color: #888;
-  border-left: 1px solid gray;
   display: inline-block;
   padding: 0.5rem;
 }
