+++ assets/css/_disruption_form.scss
@@ -1,0 +1,10 @@
+.m-disruption-form__header-id {
+  display: inline-block;
+  font-size: large;
+  margin-left: 2rem;
+}
+
+.m-disruption-form__header-num {
+  font-size: large;
+  font-weight: 200;
+}
