--- assets/css/_disruption_summary.scss
@@ -1,22 +1,0 @@
-.m-disruption-summary__adjustment_list {
-  padding-left: 0;
-}
-
-.m-disruption-summary__adjustment {
-  list-style: none;
-  margin-left: 0;
-  margin-top: 0.5rem;
-}
-
-.m-disruption-summary__adjustment_label {
-  font-weight: bold;
-  font-size: 1.5625rem;
-}
-
-.m-disruption-summary__adjustment_route {
-  font-style: italic;
-}
-
-.m-disruption-summary__disruption_id {
-  font-size: 1rem;
-}
