--- assets/css/_new_disruption_preview.scss
@@ -1,13 +1,0 @@
-.m-new-disruption-preview__dow_list {
-  padding-left: 0;
-}
-
-.m-new-disruption-preview__dow {
-  list-style: none;
-}
-
-.m-new-disruption-preview__day_label {
-  display: inline-block;
-  font-weight: bold;
-  width: 7rem;
-}
