+++ assets/css/app.scss
@@ -12,8 +12,7 @@
 @import "forms";
 @import "disruption_times";
 @import "disruption_details";
-@import "disruption_summary";
-@import "new_disruption_preview";
+@import "disruption_form";
 @import "disruption_diffs";
 @import "disruption_list_container";
 @import "disruption_table";
