+++ assets/package.json
@@ -12,9 +12,6 @@
     "check": "tsc --noEmit && npm run lint:check && npm run format:check"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.31",
-    "@fortawesome/free-solid-svg-icons": "^5.15.0",
-    "@fortawesome/react-fontawesome": "^0.1.10",
     "@fullcalendar/daygrid": "^5.3.2",
     "@fullcalendar/react": "^5.3.1",
     "bootstrap": "^4.5.2",
